requirements.txt
overrides
parsimonious
allennlp
absl-py
nni
pandas
mysql-connector-python